refactor: src/types/ をfeature層とshared層に移動
