--- src/types/index.ts
@@ -1,7 +1,0 @@
-export * from './article';
-export * from './user';
-export * from './comment';
-export * from './stock';
-export * from './like';
-export * from './notification';
-export * from './api';
